requirements.txt
pywinrm==0.3.0